Enable Google OAuth2 login from the command line.
You need to pass the following arguments to enable it:

  --oauth2-client-secrets-file <filename>
  --oauth2-user-credentials-file <filename>

The client-secrets file is assumed to be in the format of the
client_secrets.json that you can download when you register your
app with Google.

The user credentials file is, provisionally, a JSON object with
the following attributes:

  username (e.g. aranhoide@gmail.com)
  access_token (optional)
  refresh_token

Don't rely on this just yet, though.  I may change it to whatever
format the Google OAuth2 Java library uses, if that makes sense.

At this moment you still need to provide a local password, either
through the --password-file argument or in the prompt.  I'm not
sure that this is needed anymore since the client-secrets and the
oauth2-user-credentials provide all the data you need to log in.