Add command line option to specify give mode proxy port.
This is done so we can pick the port at the time we launch EC2
instances, so we can set up the AWS _security group_ (~firewall)
rules at instance launch time.

This also allows us to keep the port constant across Lantern runs
so Lantern is easier to find for get mode users without access to
Google Talk.