correctly maintain "setupComplete" state
Switching from Give to Get Mode for the first time means the user has never
completed the Get Mode-specific modals, and should be taken back to that point
in the setup sequence. The "setupComplete" field should be toggled from true
to false back to true as this is happening. The UI uses it to determine
correct behavior of the modals.