fix systemProxy == false bug in systemProxy modal
According to http://docs.angularjs.org/api/ng.$http you cannot pass a boolean
as data, must be string or object. The fact that it was working when we were
passing "true" was a happy accident. /interaction/continue on the systemProxy
modal now passes a payload like {"path": "/settings/systemProxy", "value": true|false}
which actually gets sent when value is false.