Always append StartupWMClass line to lantern.desktop.
The deleted check was causing the StartupWMClass to fall out of
sync with Lantern's random URL path whenever the latter was reset
(for example, after a user wipes the .lantern folder.)

When there are multiple StartupWMClass lines in a .desktop file,
the last one applies.  Thus, it is safe to always append.

Resetting the random URL path should be a rare enough event that
the buildup of StartupWMClass lines in lantern.desktop shouldn't
be a problem.