improve google logout logic, disable private browsing in Cocoa WebView
Only use the oauth url that first logs the user out of Google after the user
has already tried logging in with a user that's not in the private beta. This
is necessary for #981 to be able to save the user from having to enter her
password if she's already logged in to google from her default browser. This
also saves an unnecessary initial logout request and redirect response.

Also disable private browsing in the Cocoa WebView, so that it shares Safari's
cookie jar. This allows lantern-ui's Google Analytics to properly track the
user across runs, and saves users on OS X from having to type in their password
if they're already logged in to Google from Safari. closes #995.