state machine to start/stop flashlight
Listens to IP, disconnect, and changes between give and get mode to
make sure the flashlight server runs if and only if we have an IP
and we are in give mode.

This is only the triggering logic.  Actual starting, registering,
stopping, and unregistering the local flashlight server are not
done yet.