program to find and delete test CloudFront distributions
Deleting CloudFront distributions is a multi-step process that may
take up to half an hour:

http://docs.aws.amazon.com/AmazonCloudFront/latest/APIReference/DeleteDistribution.html

This will be uploaded to a cloud VM (e.g., the cloudmaster) and run
as a cron job to clean up CloudFlare distributions created by
cfr_test.go.

In the future this can be used to clean up other distributions that
we mark as unused.