make golang output files optional and configurable
I want to run genconfig in cloudmasters, where I don't need the go output
files.  I don't want to introduce a dependency on gofmt (and hence on a
go installation) nor make a dummy directory structure to receive the unneeded
golang-formatted output files.  Also, it's slightly cleaner not to hardcode
these paths in the go source.